requirements.txt
# Core Dependencies
qdrant-client==1.11.3
sentence-transformers==2.2.2
torch==2.1.0

# Chemistry
rdkit==2023.9.1
pubchempy==1.0.4

# Bio Data
biopython==1.81

# Web Framework
streamlit==1.28.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.0

# HTTP & APIs
requests==2.31.0

# LLM Integration
groq==0.4.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1